requirements.txt
# Tooling prerequisites (install manually, not via pip)
# - Terraform >= 1.1.7
# - Azure CLI >= 2.52.0

# Python packages needed for ingestion scripts
azure-storage-blob>=12.19.0
requests>=2.31.0
pandas>=2.2.0
beautifulsoup4>=4.12.0
openpyxl>=3.1.5
sqlalchemy>=2.0.19
pyodbc>=4.0.39
